refactor: Router 구조 리팩토링
