+++ src/modules/compare/routes.js
@@ -1,0 +1,9 @@
+import ComparePage from '@/modules/compare/pages/ComparePage.vue'
+
+export const compareRoutes = [
+  {
+    path: '/compare',
+    name: 'ComparePage',
+    component: ComparePage
+  }
+]
