+++ src/modules/home/routes.js
@@ -1,0 +1,10 @@
+import HomePage from '@/modules/home/pages/HomePage.vue'
+
+export const homeRoutes = [
+  {
+    path: '/',
+    name: 'home',
+    component: HomePage,
+    meta: { requiresAuth: false }
+  }
+]
